fix bug

+++ src/Helper/VendorRequest.php
@@ -51,11 +51,11 @@ class VendorRequest
      * VendorRequest constructor.
      * @param RequestInterface $request
      * @param string $vendorCode
-     * @param float $execStart
+     * @param int $execStart
      * @param string $response
      * @param string $responseOther
      */
-    public function __construct(RequestInterface $request, string $vendorCode, float $execStart, string $response, string $responseOther = '')
+    public function __construct(RequestInterface $request, string $vendorCode, int $execStart, string $response, string $responseOther = '')
     {
         $this->vendorCode = $vendorCode;
         $this->requestPath = $request->path();
